requirements.txt
py-cord
yaml
discord-py-slash-command